--- src/converters/index.ts
@@ -1,1 +1,0 @@
-export * from './manual_task_converter';
